The README.md code breaks with store undefined
`render()` is called from  `dispatch({ type: '@@INIT' })` on line 189 of this README.md or line 33 in the project code of `reducer.js` (`container.textContent = store.getState.count`), before store is defined, so you get a undefined `store` error. So you need to place a `if` statement to catch it if it's empty. Also `getState` was changed to invoke, `getState()`. So the complete change is just `if(!!store) container.textContent = store.getState().count;`